move to snapping/move.tsx
