+++ src/app/views/canvas/snapping/move.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-import { Specification, Prototypes } from "../../../core";
-import { SnappingSession, SnappingAction } from "./snapping/common";
+import { Specification, Prototypes } from "../../../../core";
+import { SnappingSession, SnappingAction } from "./common";
 
 export class MoveSnappingSession extends SnappingSession<void> {
   constructor(handle: Prototypes.Handles.Description) {
